move : profileBox 이동
